// src/app/components/models/Profile.ts
export interface profile{

    cname : string
    ccontact: string
    phone: string
    second_phone: string
    street: string
    senderaccount: string
    paymnetaccount: string
    shipmentcredential: string
    preferedservice: string
    secondservice: string
    packaging: string
    id: string
    
}